requirements.txt
boto3==1.13.3
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1
PyMySQL==0.9.3
pandas==0.22.0
sklearn==0.0
xgboost==0.90
pytest==5.3.5